% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiting_time.R
\name{compute_waiting_time}
\alias{compute_waiting_time}
\title{Compute the Minimum Waiting Time for Each Query}
\usage{
compute_waiting_time(routes_df)
}
\arguments{
\item{routes_df}{A tidy tibble containing connection details.}
}
\value{
A tibble containing only one row per query
(the next available departure)
with \code{waiting_time_min} added.
}
\description{
Calculates the waiting time (in minutes) between the requested
timetable query time and the actual departure time of each connection.
Filters out connections that departed before the query time, then keeps
only the connection with the smallest non-negative waiting time for each
unique origin-destination-query_time combination.
}
