% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json_data.R
\name{parse_route_data}
\alias{parse_route_data}
\title{Parse API JSON Response into a Tidy Table}
\usage{
parse_route_data(api_response)
}
\arguments{
\item{api_response}{A list containing the parsed JSON
(e.g., loaded from a cached .rds file).}
}
\value{
A tidy tibble containing connection details.
}
\description{
Extracts connection details from the raw JSON files
and formats them into a tidy data frame.
}
