% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{plot_accessibility_map}
\alias{plot_accessibility_map}
\title{Creates waiting-time accessibility map}
\usage{
plot_accessibility_map(waiting_times_data, station_data)
}
\arguments{
\item{waiting_times_data}{A tidy data frame containing computed waiting times
per destination (output of \code{summarise_waiting()}).}

\item{station_data}{A tidy data frame containing regional station coordinates
and population (output of \code{read_station_data()}).}
}
\value{
A ggplot object representing the accessibility map
}
\description{
Generates a regional accessibility map overlaying destination cities onto a
Swiss base map. Point sizes reflect population, and colors reflect median
waiting time.
}
